Fixing all endpoints and controllers

+++ controllers/user.go
@@ -1,39 +1,38 @@
 package controllers
 
 import (
+	final_project "github.com/ZhanserikKalmukhambet/Go_Final_Project"
 	"github.com/ZhanserikKalmukhambet/Go_Final_Project/initializers"
+	"github.com/ZhanserikKalmukhambet/Go_Final_Project/middleware"
 	"github.com/ZhanserikKalmukhambet/Go_Final_Project/models"
 	"github.com/gin-gonic/gin"
 	"net/http"
-	"strconv"
 )
 
-func ListOfUser(c *gin.Context) {
-	var user []models.User
+func ListOfUsers(c *gin.Context) {
+	var users []models.User
 
-	initializers.DB.Find(&user)
-	if len(user) == 0 {
+	initializers.DB.Find(&users)
+	if len(users) == 0 {
 		c.JSON(http.StatusBadRequest, gin.H{"error": "Couldn't retrieve data!"})
 		return
 	}
 
-	c.JSON(http.StatusOK, gin.H{"data": user})
-}
+	isAuth := final_project.IsAuthorizedOrReadOnly(c)
 
-func CreateUser(c *gin.Context) {
-	var user models.User
-
-	if err := c.ShouldBindJSON(&user); err != nil {
-		c.JSON(http.StatusBadRequest, gin.H{"error": err.Error()})
+	if !isAuth {
+		c.JSON(http.StatusUnauthorized, gin.H{"error": "You are unauthorized."})
 		return
 	}
 
-	if err := initializers.DB.Create(&user).Error; err != nil {
-		c.JSON(http.StatusBadRequest, gin.H{"error": err.Error()})
+	isAdmin := final_project.IsAdmin(c)
+
+	if !isAdmin {
+		c.JSON(http.StatusForbidden, gin.H{"error": "You are not Admin."})
 		return
 	}
 
-	c.JSON(http.StatusOK, gin.H{"data": user})
+	c.JSON(http.StatusOK, gin.H{"data": users})
 }
 
 func GetUserByID(c *gin.Context) {
@@ -43,21 +42,21 @@ func GetUserByID(c *gin.Context) {
 		c.JSON(http.StatusBadRequest, gin.H{"error": "Record not found!"})
 		return
 	}
-	c.JSON(http.StatusOK, gin.H{"data": user})
-}
 
-func UpdateUserByID(c *gin.Context) {
-	var user models.User
-	if err := initializers.DB.Where("id = ?", c.Param("id")).First(&user).Error; err != nil {
-		c.JSON(http.StatusBadRequest, gin.H{"error": "Record not found!"})
+	isAuth := final_project.IsAuthorizedOrReadOnly(c)
+
+	if !isAuth {
+		c.JSON(http.StatusUnauthorized, gin.H{"error": "You are unauthorized."})
 		return
 	}
 
-	if err := c.ShouldBindJSON(&user); err != nil {
-		c.JSON(http.StatusBadRequest, gin.H{"error": err.Error()})
+	isAdmin := final_project.IsAdmin(c)
+
+	if !isAdmin {
+		c.JSON(http.StatusForbidden, gin.H{"error": "You are not Admin."})
 		return
 	}
-	initializers.DB.Save(&user)
+
 	c.JSON(http.StatusOK, gin.H{"data": user})
 }
 
@@ -67,26 +66,88 @@ func DeleteUserByID(c *gin.Context) {
 		c.JSON(http.StatusBadRequest, gin.H{"error": "Record not found!"})
 		return
 	}
+
+	isAuth := final_project.IsAuthorizedOrReadOnly(c)
+
+	if !isAuth {
+		c.JSON(http.StatusUnauthorized, gin.H{"error": "You are unauthorized."})
+		return
+	}
+
+	isAdmin := final_project.IsAdmin(c)
+
+	if !isAdmin {
+		c.JSON(http.StatusForbidden, gin.H{"error": "You are not Admin."})
+		return
+	}
+
 	initializers.DB.Delete(&user)
 	c.JSON(http.StatusOK, gin.H{"data": "Record deleted!"})
 }
 
-func UserChats(c *gin.Context) {
-	UserID, err := strconv.Atoi(c.Param("id"))
-
-	if err != nil {
-		panic(err)
-	}
+func ListOfUserChats(c *gin.Context) {
+	UserID := int(middleware.GetPayloadFromToken(c)["userID"].(float64))
 
 	var chats []models.Chat
-	var myChats []models.Chat
-	initializers.DB.Find(&chats)
 
-	for _, val := range chats {
-		if val.UserID == UserID {
-			myChats = append(myChats, val)
-		}
+	if err := initializers.DB.Where("user_id = ?", UserID).First(&chats).Error; err != nil {
+		c.JSON(http.StatusBadRequest, gin.H{"error": "Record not found!"})
+		return
 	}
 
-	c.JSON(http.StatusOK, myChats)
+	c.JSON(http.StatusOK, chats)
+}
+
+func ListOfUserFavoriteItems(c *gin.Context) {
+	if final_project.IsAuthorizedOrReadOnly(c) == false {
+		c.JSON(http.StatusUnauthorized, gin.H{"error": "Unauthorized!"})
+		return
+	}
+
+	loggedUser := middleware.GetPayloadFromToken(c)
+	if loggedUser == nil {
+		c.JSON(http.StatusBadRequest, gin.H{"error": "Can't load data from token!"})
+		return
+	}
+
+	var favoriteItems []models.FavoriteItem
+
+	if err := initializers.DB.Where("user_id = ?", int(loggedUser["userID"].(float64))).First(&favoriteItems).Error; err != nil {
+		c.JSON(http.StatusBadRequest, gin.H{"error": "Record not found!"})
+		return
+	}
+
+	if len(favoriteItems) == 0 {
+		c.JSON(http.StatusBadRequest, gin.H{"error": "Couldn't retrieve data!"})
+		return
+	}
+
+	c.JSON(http.StatusOK, gin.H{"data": favoriteItems})
+}
+
+func ListOfUserAdvertisements(c *gin.Context) {
+	if final_project.IsAuthorizedOrReadOnly(c) == false {
+		c.JSON(http.StatusUnauthorized, gin.H{"error": "Unauthorized!"})
+		return
+	}
+
+	loggedUser := middleware.GetPayloadFromToken(c)
+	if loggedUser == nil {
+		c.JSON(http.StatusBadRequest, gin.H{"error": "Can't load data from token!"})
+		return
+	}
+
+	var advertisements []models.Advertisement
+
+	if err := initializers.DB.Where("user_id = ?", int(loggedUser["userID"].(float64))).First(&advertisements).Error; err != nil {
+		c.JSON(http.StatusBadRequest, gin.H{"error": "Record not found!"})
+		return
+	}
+
+	if len(advertisements) == 0 {
+		c.JSON(http.StatusBadRequest, gin.H{"error": "Couldn't retrieve data!"})
+		return
+	}
+
+	c.JSON(http.StatusOK, gin.H{"data": advertisements})
 }
